requirements.txt
streamlit
openai==0.28.0
pyautogen==0.1.14